openapi-generator/prism-agent-client/pyproject.toml
[tool.poetry]
name = "prism-agent-client"
version = "1.0.0"
description = "A client library for accessing Prism Agent"

authors = []

readme = "README.md"
packages = [
    {include = "prism_agent_client"},
]
include = ["CHANGELOG.md", "prism_agent_client/py.typed"]

[tool.poetry.dependencies]
python = "^3.8"
httpx = ">=0.15.4,<0.25.0"
attrs = ">=21.3.0"
python-dateutil = "^2.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target_version = ['py38', 'py39', 'py310', 'py311']
exclude = '''
(
  /(
    | \.git
    | \.venv
    | \.mypy_cache
  )/
)
'''

[tool.isort]
line_length = 120
profile = "black"
